Get a video playing in response to the add command
The code isn't clean, but with this commit, I can see audio
playing through the bot in discord. This is done manually with
an mp3 file using docs from the guide. Next steps will
involve figuring out how to get ytdl-core to create a stream
that can be played via discord